refactor: move chat to separate module
