+++ client/src/components/GameResults/GameResults.module.scss
@@ -34,32 +34,3 @@
   align-items: center;
   gap: 2rem;
 }
-.chart {
-  width: 35%;
-  max-height: 100%;
-  position: relative;
-}
-.chart__info {
-  position: absolute;
-  max-width: 100%;
-  max-height: 100%;
-  aspect-ratio: 1/1;
-  top: .5rem;
-  font-size: 1.3rem;
-  padding: 1.5rem;
-  border-radius: 50%;
-  display: flex;
-  gap: 10px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-.canvas {
-  height: 100%;
-  width: 100%;
-}
-.level {
-  font-weight: bold;
-  font-size: 1.5rem;
-}
