+++ client/src/components/GameResults/GameResults.tsx
@@ -1,11 +1,9 @@
 import { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { Doughnut } from "react-chartjs-2";
-import { Chart as ChartJS, ArcElement, Tooltip, Legend } from "chart.js";
 import { Button } from "../UI";
 import { ResultItem } from "./ResultItem.tsx/ResultItem";
 import styles from "./GameResults.module.scss";
-import { NEURONS_ON_LEVEL } from "../../constants/constants";
+import { NeuronsChart } from "./ResultItem.tsx/NeuronsChart/NeuronsChart";
 
 interface GameResultsProps {
   correctAnswers: number;
@@ -19,8 +17,6 @@ interface Result {
   title: string;
   value: string | number;
 }
-
-ChartJS.register(ArcElement, Tooltip, Legend);
 
 export const GameResults: FC<GameResultsProps> = ({
   correctAnswers = 0,
@@ -52,24 +48,6 @@ export const GameResults: FC<GameResultsProps> = ({
 
   // TO DO: Get number of neurons gained by user from server and save it to userNeurons instead of random number below:
   const userNeurons = 3228;
-  const currentLevelNeurons = userNeurons % NEURONS_ON_LEVEL;
-  const level = Math.ceil(userNeurons / NEURONS_ON_LEVEL);
-  const neuronsToTheNextLevel = NEURONS_ON_LEVEL - currentLevelNeurons;
-  const maxNeuronsOnLevel = NEURONS_ON_LEVEL * level;
-
-  const data = {
-    labels: [],
-    datasets: [
-      {
-        label: "Neurons",
-        data: [currentLevelNeurons, neuronsToTheNextLevel],
-        backgroundColor: ["rgba(96, 57, 170, 1)", "rgba(222, 222, 222, 1)"],
-        borderColor: ["#9e9e9e"],
-        borderWidth: 1,
-        cutout: "90%",
-      },
-    ],
-  };
 
   return (
     <div className={styles.wrapper}>
@@ -86,17 +64,7 @@ export const GameResults: FC<GameResultsProps> = ({
             );
           })}
         </div>
-        <div className={styles.chart}>
-          <Doughnut className={styles.canvas} data={data} />
-          <div className={styles.chart__info}>
-            <span className={styles.level}>
-              {level} {t("gameResults.level")}
-            </span>
-            <span>
-              {userNeurons} {t("gameResults.from")} {maxNeuronsOnLevel} {t("gameResults.neurons")}
-            </span>
-          </div>
-        </div>
+        <NeuronsChart userNeurons={userNeurons} />
       </div>
       <Button onClick={newGameHandler}>{t("gameResults.playAgain")}</Button>
     </div>
